mudança de double para long para que possa comportar numeros grandes

+++ src/control/Facada.java
@@ -7,14 +7,14 @@ public interface Facada {
 	 * @param baseRaiz = raiz que o numero será elevado.
 	 * @return Raiz
 	 */
-	public float calcularRaiz(float numero, float baseRaiz);
+	public long calcularRaiz(float numero, float baseRaiz);
 	
 	/**Retorna um numero elevado em uma potencia.
 	 * @param numero = numero que será elevado.
 	 * @param basePotencia = potencia que elevará o numero.
 	 * @return Potencia
 	 */
-	public float calcularPotencia(float numero, float basePotencia);
+	public long calcularPotencia(float numero, float basePotencia);
 	
 	/**
 	 * @param a = incognita ao quadrado (X²).
@@ -22,7 +22,7 @@ public interface Facada {
 	 * @param c = termo conhecido (numero).
 	 * @return a raiz, ou as raizes de umas eqaução do 2º Grau.
 	 */
-	public float[] calcularRaizEquacao(float a, float b, float c); 
+	public double[] calcularRaizEquacao(float a, float b, float c); 
 	
 	/**
 	 * @param a = incognita ao quadrado (X²).
